+++ src/app/admin/users/manage-sign/manage-sign.component.ts
@@ -1,81 +1,110 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, OnDestroy } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
 import { SignsService } from '../../../shared/signs.service';
 import { UserSignService } from '../../../shared/user-sign.service';
 import { NotifierService } from 'angular-notifier';
 
-// import { FormGroup, FormBuilder, FormArray, ReactiveFormsModule } from '@angular/forms';
+// import { Subject } from 'rxjs';
+import { FormBuilder, FormGroup, FormArray } from '@angular/forms';
+import { ConfirmationService } from 'primeng/api';
+import { NgxSpinnerService } from 'ngx-spinner';
+
 @Component({
   selector: 'app-manage-sign',
   templateUrl: './manage-sign.component.html',
   styleUrls: ['./manage-sign.component.css']
 })
-export class ManageSignComponent implements OnInit {
+export class ManageSignComponent implements OnInit, OnDestroy {
   userid: any;
   user_name: string;
   user_role: string;
-  // manageSign: FormGroup;
-  // orderForm: FormGroup;
-  // items: FormArray;
+  signForm: FormGroup;
+  signs: FormArray;
   constructor(
-    // private formBuilder: FormBuilder, private _fb: FormBuilder,
+    private _fb: FormBuilder,
     private route: ActivatedRoute, private service: SignsService,
-    private usservice: UserSignService, private notifier: NotifierService) { }
+    private usservice: UserSignService, private notifier: NotifierService,
+    private confirmationService: ConfirmationService, private spinner: NgxSpinnerService) { }
   allSign: any = [];
   fieldArray: any = [];
   newAttribute: any = {};
-
+  addNewSign: any;
   firstField = true;
   firstFieldName = 'MacDonalds-Irvine';
   isEditItems: boolean;
   isAddNew: boolean;
   oldChoice: Number = 0;
   newChoice: Number;
+  // dtTrigger = new Subject();
+  // dtOptions: DataTables.Settings = {};
   ngOnInit() {
-    // this.user_name = localStorage.getItem('name');
-    // this.user_role = (localStorage.getItem('authorities')).replace('ROLE_', '');
-
-    // this.manageSign = this._fb.group(
-    //   {
-    //     itemRows: this._fb.array([this.initItemRows()])
-    //   }
-    //   // {
-    //   //   customerName: '',
-    //   //   email: '',
-    //   //   items: this._fb.array([this.initItemRows()])
-    //   // }
-    // );
-
-    // this.orderForm = this.formBuilder.group({
-    //   customerName: '',
-    //   email: '',
-    //   items: this.formBuilder.array([this.createItem()])
-    // });
-
+    // this.dtOptions = {
+    //   pagingType: 'full_numbers',
+    //   pageLength: 10,
+    //   order: [0, 'desc']
+    // };
+    this.signForm = this._fb.group({
+      itemRows: this._fb.array([this.initItemRows()]) // here
+    });
     this.route.params.subscribe(params => {
       this.userid = params['id'];
       this.getSignByUser();
     });
+    this.getAllSigns();
+  }
 
+  initItemRows() {
+    return this._fb.group({
+      // list all your form controls here, which belongs to your form array
+      itemname: ['']
+    });
+  }
+
+  // addNewRow() {
+  //   // control refers to your formarray
+  //   const control = <FormArray>this.signForm.controls['itemRows'];
+  //   // add new formgroup
+  //   control.push(this.initItemRows());
+  // }
+
+  // deleteRow(index: number) {
+  //   // control refers to your formarray
+  //   const control = <FormArray>this.signForm.controls['itemRows'];
+  //   // remove the chosen row
+  //   control.removeAt(index);
+  // }
+
+
+  getAllSigns() {
+    this.spinner.show();
     this.service.getAllSigns().subscribe(res => {
       this.allSign = res;
-      // // console.log(res);
+      // console.log(res);
+      this.spinner.hide();
+    }, error => {
+      console.log(error);
+      this.spinner.hide();
     });
-
   }
-
   getSignByUser() {
+    this.spinner.show();
     this.usservice.getSignByUserId_admin(this.userid).subscribe(res => {
       this.fieldArray = res;
-      // console.log(res);
+      // this.dtTrigger.next();
+      // console.log('fieldArray', res);
+      this.spinner.hide();
+    }, error => {
+      console.log(error);
+      this.spinner.hide();
     });
+
   }
 
-  onfocusSign(id) {
-    // this.notifier.notify('info', event.target.value);
-    // console.log('old : ' + id);
-    this.oldChoice = id;
-  }
+  // onfocusSign(id) {
+  //   // this.notifier.notify('info', event.target.value);
+  //   // console.log('old : ' + id);
+  //   this.oldChoice = id;
+  // }
 
   onchangeSign(id) {
     // this.notifier.notify('info', event.target.value);
@@ -97,7 +126,7 @@ export class ManageSignComponent implements OnInit {
     this.usservice.addUserSign(this.newChoice, this.userid)
       .subscribe(res => {
         this.notifier.notify('success', 'Updated Successfully');
-        this.onEditCloseItems();
+        // this.onEditCloseItems();
       }, error => {
         if (error.status === 500) {
           this.notifier.notify('error', 'Already Exist.');
@@ -105,24 +134,29 @@ export class ManageSignComponent implements OnInit {
           // console.log(error);
         }
       });
-    this.oldChoice = id;
+    // this.oldChoice = id;
 
   }
 
-  addUserSign(id) {
-    // console.log(id.toString().substr(3, 1));
-    this.usservice.addUserSign(id.toString().substr(3, 1), this.userid)
+  addUserSign() {
+    // console.log(this.addNewSign + ' - ' + this.userid);
+    this.spinner.show();
+    this.usservice.addUserSign(this.addNewSign, this.userid)
       .subscribe(res => {
         this.notifier.notify('success', 'Added Successfully');
-        this.onEditCloseItems();
+        this.spinner.hide();
+        this.getSignByUser();
+        // this.onEditCloseItems();
       }, error => {
         if (error.status === 500) {
-          this.notifier.notify('error', 'Already Exist.');
+          this.notifier.notify('error', error.error.message);
+          this.spinner.hide();
         } else {
-          // console.log(error);
+          console.log(error);
+          this.spinner.hide();
         }
       });
-    this.addFieldValue();
+    // this.addFieldValue();
   }
 
   addFieldValue() {
@@ -132,22 +166,47 @@ export class ManageSignComponent implements OnInit {
   }
 
   deleteFieldValue(id) {
-    this.usservice.deleteUserSign(id).subscribe(res => {
-      // this.fieldArray.splice(index, 1);
-      this.notifier.notify('warning', res.toString());
-      this.getSignByUser();
-    }, error => {
-      // this.notifier.notify('error', 'Something Wrong');
-      this.getSignByUser();
-    }
-    );
+
+    this.confirmationService.confirm({
+      message: 'Do you want to delete this record?',
+      header: 'Delete Confirmation',
+      icon: 'pi pi-info-circle',
+      accept: () => {
+        this.spinner.show();
+        this.usservice.deleteUserSign(id).subscribe(res => {
+          // this.fieldArray.splice(index, 1);
+          this.notifier.notify('warning', res.toString());
+          this.spinner.hide();
+          this.getSignByUser();
+        }, error => {
+          // this.notifier.notify('error', 'Something Wrong');
+          // this.getSignByUser();
+          if (error.status === 200) {
+            this.notifier.notify('warning', error.error.text);
+            this.getSignByUser();
+          } else {
+            console.log(error);
+          }
+          this.spinner.hide();
+        }
+        );
+      },
+      reject: () => {
+        this.notifier.notify('info', 'Request Rejected For Delete');
+      }
+    });
+
   }
 
-  onEditCloseItems() {
-    this.isEditItems = !this.isEditItems;
-    this.getSignByUser();
-  }
+  // onEditCloseItems() {
+  //   this.isEditItems = !this.isEditItems;
+  //   this.getSignByUser();
+  // }
 
+  ngOnDestroy(): void {
+    // Do not forget to unsubscribe the event
+    // this.dtTrigger.unsubscribe();
+  }
 
   // // *** DYNAMIC GENERATES CONTROLS ** //
